core/requirements.txt
pyyaml==5.1.2
requests==2.22.0
